+++ dask/array/core.py
@@ -3488,8 +3488,15 @@ def _vindex(x, *indexes):
     x = x[nonfancy_indexes]
 
     array_indexes = {}
+    dask_array_indexes = {}
     for i, (ind, size) in enumerate(zip(reduced_indexes, x.shape)):
-        if not isinstance(ind, slice):
+        if isinstance(ind, Array):
+            if ind.dtype.kind == 'b':
+                raise IndexError('vindex does not support indexing with '
+                                 'boolean arrays')
+            ind %= size
+            dask_array_indexes[i] = ind
+        elif not isinstance(ind, slice):
             ind = np.array(ind, copy=True)
             if ind.dtype.kind == 'b':
                 raise IndexError('vindex does not support indexing with '
@@ -3503,6 +3510,9 @@ def _vindex(x, *indexes):
 
     if array_indexes:
         x = _vindex_array(x, array_indexes)
+
+    if dask_array_indexes:
+        x = _vindex_dask_arrays(x, dask_array_indexes)
 
     return x
 
@@ -3585,6 +3595,90 @@ def _vindex_array(x, dict_indexes):
         tuple(map(sum, chunks)), chunks=chunks, dtype=x.dtype, name=out_name
     )
     return result_1d.reshape(broadcast_shape + result_1d.shape[1:])
+
+
+def _vindex_dask_arrays(x, dict_indexes):
+    """ Point wise indexing with Dask Arrays
+
+    Makes two passes of the data. First pass with all static values: integers,
+    slices, lists, and NumPy Arrays. Second pass (done in this function)
+    handles Dask Arrays using ``atop``. The strategy there involves comparing
+    Dask Array indices with ``arange`` in chunks to determine the relative
+    indexing needed to extract values for that chunk. This strategy works just
+    as well for Dask Arrays of known and unknown length.
+    """
+
+    from .creation import arange
+
+    # Construct a tuple of indices for selection.
+    # Also collect axes that are part of the result.
+    #
+    # * arr_inds -- used for second pass
+    # * out_axes -- track where each
+    #
+    arr_inds = []
+    out_axes = []
+    for i in range(x.ndim):
+        try:
+            ind = dict_indexes[i]
+        except KeyError:
+            arr_inds.append(slice(None))
+            out_axes.append(len(arr_inds))
+        else:
+            arr_inds.append(ind)
+            if 0 not in out_axes:
+                out_axes.append(0)
+    arr_inds = tuple(arr_inds)
+    out_axes = tuple(out_axes)
+
+    # Find indices for all points in x to aid in selection within chunks
+    x_inds = tuple(
+        arange(s, chunks=c) for s, c in zip(x.shape, x.chunks)
+    )
+
+    # Use global indices for reference in applying selected indices relative
+    # to each chunk. For static values applied in the first pass, pass along
+    # a dummy scalar value to notify the chunk should be kept unchanged.
+    idx_vals = atop(
+        _vindex_dask_arrays_chunk,
+        out_axes,
+        x, tuple(range(1, 1 + x.ndim)),
+        *concat([
+            (e_0, (0,), e_i, (i,)) if isinstance(e_0, Array) else
+            (e_0, None, e_i, (i,))
+            for i, (e_0, e_i) in enumerate(zip(arr_inds, x_inds), start=1)
+        ]),
+        dtype=x.dtype,
+        concatenate=True
+    )
+
+    return idx_vals
+
+
+def _vindex_dask_arrays_chunk(xc, *args):
+    """ Point wise indexing with Dask Arrays within each chunk
+
+    Uses the global indices provided to this chunk both for reference (as to
+    where this chunk lies in the larger Dask Array) and for selection (user
+    provided, but still in the absolute reference frame). Find the selection
+    in chunk relative indices and apply it.
+    """
+
+    # Unpack interwoven reference indices and selection indices.
+    inds = args[::2]
+    x_inds = args[1::2]
+
+    xc_inds = []
+    for i, (e_0, e_i) in enumerate(zip(inds, x_inds)):
+        if isinstance(e_0, np.ndarray):
+            xc_inds.append(np.searchsorted(e_i, e_0[e_0 <= e_i[-1]]))
+        else:
+            xc_inds.append(e_0)
+    xc_inds = tuple(xc_inds)
+
+    vals = xc[xc_inds]
+
+    return vals
 
 
 def _get_axis(indexes):